backend/requirements.txt
Django~=5.2.6
psycopg2==2.9.10
djangorestframework~=3.16.1
djangorestframework-simplejwt~=5.5.1
django-cors-headers~=4.8.0
python-decouple==3.8
cryptography~=46.0.1
python-jose~=3.5.0
requests~=2.32.5